core: change speed option/property to double
The --speed option and the speed property used float. Change them to
double.

Change the commands that manipulate the property (speed_mult/add) to
double as well. Since the cycle command shares code with the add
command, we change that as well.

The reason for this change is that this allows better control over
speed, such as stepping by semitones. Using floats is also just plain
unnecessary.

+++ audio/filter/af_scaletempo.c
@@ -444,11 +444,11 @@ static int control(struct af_instance* af, int cmd, void* arg)
       if (s->speed_pitch) {
         break;
       }
-      s->speed = *(float*)arg;
+      s->speed = *(double *)arg;
       s->scale = s->speed * s->scale_nominal;
     } else {
       if (s->speed_pitch) {
-        s->speed = 1 / *(float*)arg;
+        s->speed = 1 / *(double *)arg;
         s->scale = s->speed * s->scale_nominal;
         break;
       }
